defconfig: Disable CMDLINE debugging configs and remove LTO.
We're past the point of needing those on G5 and V20. G6 hasn't
been tested as much, but it's very unlikely that it might need
this kind of debugging, given most of its latest issues were
related to USB.

LTO references on that config were also removed, as it would need
a major overhaul on the kernel tree to be enabled, and most of
the issues come from updated assembly instructions and backports,
which is just too much work for something that might squeeze a few
more percent of performance at most.